+++ background-tasks/src/export-template/copyTemplateSources.js
@@ -5,10 +5,14 @@ import { getLogger } from 'bbcat-orchestration-builder-logging';
 
 const logger = getLogger('export-template');
 
-const templateSourceDir = path.dirname(require.resolve('@bbc/bbcat-orchestration-template/package.json')).replace('app.asar', 'app.asar.unpacked');
+const defaultTemplateSourceDir = path.dirname(require.resolve('@bbc/audio-orchestration-template/package.json'))
+  .replace('app.asar', 'app.asar.unpacked');
 
 const copyTemplateSources = (args) => {
-  const { outputDir } = args;
+  const { outputDir, settings = {} } = args;
+  const { customTemplatePath } = settings;
+
+  const templateSourceDir = customTemplatePath || defaultTemplateSourceDir;
 
   logger.debug(`copying template source files from ${templateSourceDir} to ${outputDir}`);
 
@@ -20,6 +24,11 @@ const copyTemplateSources = (args) => {
         'node_modules',
         'audio',
       ].includes(path.basename(file))));
+
+      if (!filesToCopy.map(file => path.basename(file)).includes('dist')) {
+        logger.error(`no 'dist' directory in template path ${templateSourceDir}`);
+        throw new Error('Template source directory does not include expected \'dist\' directory, so cannot be used.');
+      }
 
       return mapSeries(
         filesToCopy,