+++ react-frontend/src/Components/Project/Audio/Sequence/BehaviourParameters/BehaviourParameter.jsx
@@ -76,12 +76,12 @@ BehaviourParameter.propTypes = {
     PropTypes.arrayOf(PropTypes.any),
   ]).isRequired,
   sequencesList: PropTypes.arrayOf(PropTypes.shape({
-    sequenceId: PropTypes.String,
-    name: PropTypes.String,
+    sequenceId: PropTypes.string,
+    name: PropTypes.string,
   })).isRequired,
   objectsList: PropTypes.arrayOf(PropTypes.shape({
-    objectNumber: PropTypes.String,
-    label: PropTypes.String,
+    objectNumber: PropTypes.oneOfType(PropTypes.number, PropTypes.string),
+    label: PropTypes.string,
   })).isRequired,
   controls: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
   controlId: PropTypes.string,