+++ react-frontend/src/Components/Project/Audio/Sequence/BehaviourParameters/index.jsx
@@ -43,12 +43,12 @@ BehaviourParameters.propTypes = {
   values: PropTypes.shape({}).isRequired,
   onChange: PropTypes.func.isRequired,
   sequencesList: PropTypes.arrayOf(PropTypes.shape({
-    sequenceId: PropTypes.String,
-    name: PropTypes.String,
+    sequenceId: PropTypes.string,
+    name: PropTypes.string,
   })).isRequired,
   objectsList: PropTypes.arrayOf(PropTypes.shape({
-    objectNumber: PropTypes.number,
-    label: PropTypes.String,
+    objectNumber: PropTypes.oneOfType(PropTypes.number, PropTypes.string),
+    label: PropTypes.string,
   })).isRequired,
   controls: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
   controlId: PropTypes.string,